feat: added roles + guards + protected routes + endpoint for role adding (onboarding)

+++ src/jobs/jobs.service.ts
@@ -1,15 +1,21 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { JobsRepository } from './jobs.repository';
 import { CreateJobDto } from './dto/create-job.dto';
 import { JobQueryDto } from './dto/job-query.dto';
 import { Job } from './entities/job.entity';
 import { UpdateJobDto } from './dto/update-job.dto';
+import { Role } from 'src/auth/roles/roles.enum';
+import { UserPayload } from 'src/auth/interfaces/user-payload.interface';
 
 @Injectable()
 export class JobsService {
   constructor(private readonly jobsRepository: JobsRepository) {}
 
-  async create(createJobDto: CreateJobDto): Promise<string> {
+  async create(createJobDto: CreateJobDto, user: UserPayload): Promise<string> {
     // Generate simple keywords for search (Free tier alternative to Algolia)
     const keywords = this.generateKeywords([
       createJobDto.title,
@@ -19,6 +25,7 @@ export class JobsService {
 
     const jobData: Job = {
       ...createJobDto,
+      createdBy: user.uid,
       keywords,
       isActive: true,
       source: 'manual',
@@ -37,8 +44,56 @@ export class JobsService {
     return this.jobsRepository.findById(id);
   }
 
-  async remove(id: string) {
-    return this.jobsRepository.delete(id);
+  async remove(id: string, user: UserPayload) {
+    const job = await this.jobsRepository.findById(id);
+
+    const isAdmin = user.roles.includes(Role.ADMIN);
+    const isOwner = job.createdBy === user.uid;
+
+    if (!isAdmin && !isOwner) {
+      throw new ForbiddenException('You can only delete your own jobs');
+    }
+
+    await this.jobsRepository.delete(id);
+  }
+
+  async update(
+    id: string,
+    updateJobDto: UpdateJobDto,
+    user: UserPayload,
+  ): Promise<Job> {
+    const existingJob = await this.findOne(id);
+
+    // Allow if Admin OR if Owner
+    const isAdmin = user.roles.includes(Role.ADMIN);
+    const isOwner = existingJob.createdBy === user.uid;
+
+    if (!isAdmin && !isOwner) {
+      throw new ForbiddenException('You can only update your own jobs');
+    }
+
+    if (!existingJob) {
+      throw new NotFoundException(`Job with ID ${id} not found`);
+    }
+    let keywords = existingJob.keywords;
+
+    if (updateJobDto.title || updateJobDto.company || updateJobDto.techStack) {
+      const title = updateJobDto.title || existingJob.title;
+      const company = updateJobDto.company || existingJob.company;
+      const techStack = updateJobDto.techStack || existingJob.techStack;
+
+      keywords = this.generateKeywords([title, company, ...techStack]);
+    }
+
+    const updatedData: Partial<Job> = {
+      ...updateJobDto,
+      keywords,
+      updatedAt: new Date(),
+    };
+
+    await this.jobsRepository.update(id, updatedData);
+    const updatedJob = await this.findOne(id);
+    return updatedJob;
   }
 
   // --- Helper: Search Tokenizer ---
@@ -48,37 +103,12 @@ export class JobsService {
     inputs.forEach((input) => {
       if (!input) return;
       // Split by space, comma, or slash
-      const words = input.toLowerCase().split(/[\s,\/]+/);
+      const words = input.toLowerCase().split(/[\s,\\/]+/);
       words.forEach((w) => {
         if (w.length > 1) set.add(w); // Ignore single chars
       });
     });
 
     return Array.from(set);
-  }
-  async update(id: string, updateJobDto: UpdateJobDto): Promise<Job> {
-    const existingJob = await this.findOne(id);
-    if (!existingJob) {
-      throw new NotFoundException(`Job with ID ${id} not found`);
-    }
-    let keywords = existingJob.keywords;
-    
-    if (updateJobDto.title || updateJobDto.company || updateJobDto.techStack) {
-      const title = updateJobDto.title || existingJob.title;
-      const company = updateJobDto.company || existingJob.company;
-      const techStack = updateJobDto.techStack || existingJob.techStack;
-      
-      keywords = this.generateKeywords([title, company, ...techStack]);
-    }
-    
-    const updatedData: Partial<Job> = {
-      ...updateJobDto,
-      keywords,
-      updatedAt: new Date(), 
-    };
-    
-    await this.jobsRepository.update(id, updatedData);
-    const updatedJob = await this.findOne(id);
-    return updatedJob;
   }
 }
